refactor(constants): move wallet keychain indices to common
